requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pyodbc>=4.0.39
tiktoken>=0.5.0
pandas>=2.0.0
numpy>=1.20.0
pytest>=7.0.0
ruff>=0.0.1
flake8>=6.0.0
mypy>=1.0.0